refactor usePosts
